refactoring
Former-commit-id: be218a1b1614eddd7050648df0bcf6331e8148b3

+++ src/spn/tests/test_parametric.py
@@ -13,7 +13,7 @@ class TestParametric(unittest.TestCase):
     def assert_correct(self, node, x, result):
         self.tested.add(type(node))
 
-        data = np.array([x]).reshape(-1, 1)
+        data = np.array([x], dtype=np.float).reshape(-1, 1)
         node.scope = [0]
         l = likelihood(node, data)
         self.assertAlmostEqual(result, l[0, 0], 5)
